pyproject.toml
[project]
name = "llmaix"
version = "0.0.25"
authors = [
  { name = "Fabian Wolf", email = "fabian.wolf2@tu-dresden.de" }
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
  "markitdown[docx, xlsx, xls, pdf]~=0.1.2",
  "click~=8.2.1",
  "pymupdf~=1.26.3",
  "pymupdf4llm==0.0.27",
  "ocrmypdf~=16.10.4",
  "openai>=1.100.1",
  "pytest>=8.4.1",
  "hatchling~=1.27.0",
  "filetype>=1.2.0"
]

[project.urls]
Homepage = "http://github.com/KatherLab/llmaixlib"
Documentation = "http://github.com/KatherLab/llmaixlib"
Issues = "http://github.com/KatherLab/llmaixlib/issues"

[build-system]
requires = ["hatchling ~= 1.27.0"]
build-backend = "hatchling.build"

[project.scripts]
llmaix = "llmaix.cli:main"

[tool.hatch.version]
path = "src/llmaix/__version__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/llmaix"]

[dependency-groups]
dev = [
  "ruff>=0.12.9",
  "ty~=0.0.1a18",
  "pytest~=8.4.1",
  "pytest-cov~=6.2.1",
  "isort~=6.0.1",
]
marker = [
  "marker-pdf[full]~=1.8.3",
]
docling = [
  "docling~=2.45.0",
  "rapidocr~=3.3.1",
  "onnxruntime",
  "rapidocr_onnxruntime"
]

# CPU-only Paddle for the paddleocr extra
paddleocr = [
  "paddleocr>=3.1.1",
  "paddlepaddle==3.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest~=8.4.1",
  "pytest-cov~=6.2.1",
  "ruff~=0.12.9",
  "ty~=0.0.1a18",
  "isort~=6.0.1",
]
marker = ["marker-pdf[full]~=1.8.3"]
docling = [
  "docling~=2.45.0",
  "rapidocr~=3.3.1",
  "onnxruntime",
  "rapidocr_onnxruntime"
]
paddleocr = [
  "paddleocr>=3.1.1",
  "paddlepaddle==3.1.0",
]
all = ["llmaix[dev,marker,docling,paddleocr]"]

[tool.pytest.ini_options]
filterwarnings = [
  "error",
  "ignore::UserWarning",
  "ignore::DeprecationWarning"
]
